anxiety-support/backend/requirements.txt
aiogram==3.7.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
